+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - `cli/ls`: Rename command `list` to `ls`
 - `cli/cat`: Rename command `show` to `cat`
 - `cli/mv`: Rename command `move` to `mv`
+- `cli/env`: Rename command `project` to `env`
 
 ### Fixed